Split GeneratorDataModels into meaningful parts
