+++ src/Controls/src/BindingSourceGen/ITypeSymbolExtensions.cs
@@ -51,4 +51,20 @@ internal static class ISymbolExtensions
 		symbol.DeclaredAccessibility == Accessibility.Public
 		|| symbol.DeclaredAccessibility == Accessibility.Internal
 		|| symbol.DeclaredAccessibility == Accessibility.ProtectedOrInternal;
+
+	internal static AccessorKind ToAccessorKind(this ISymbol symbol)
+	{
+		return symbol switch
+		{
+			IFieldSymbol _ => AccessorKind.Field,
+			IPropertySymbol _ => AccessorKind.Property,
+			_ => throw new ArgumentException("Symbol is not a field or property.", nameof(symbol))
+		};
+	}
+}
+
+public enum AccessorKind
+{
+	Field,
+	Property,
 }
